\name{dynBiplotGUI-package}
\Rdversion{1.1.0}
\alias{dynBiplotGUI-package}
\alias{dynBiplotGUI}
\docType{package}
\title{
Interactive GUI for Dynamic Biplot in R
}
\description{
A GUI to solve Dynamic Biplot, classic Biplot and HJ-Biplot. Try matrices of 2-way and 3-way. Multilanguage.
}

\details{
\tabular{ll}{
Package: \tab dynBiplotGUI\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2014-06-26\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
The dynBiplot function initialises the GUI.
\cr At present, dynBiplotGUI supports Spanish, English, French, Portuguese.
\cr dynBiplotGUI is tested on Windows 7 (32 & 64 bits), Windows XP.
}
\author{
Jaime Egido

Maintainer: Jaime Egido <jegido@usal.es>
}
\references{
  Gabriel, K. R.  (1971)
  The biplot graphical display of matrices with application to principal
  component analysis.
  \emph{Biometrika} \bold{58}, 453-467.

  Galindo, M. P. (1986)
  Una alternativa de representacion simultanea: HJ-Biplot.
  \emph{Questiio}, 10(1):13-23, 1986.
  
  Egido, J; Galindo, M. P. (2013)
  Dynamic Biplot.
  Pending publication.
}
\examples{
dynBiplot("en")
}
\keyword{package}
\keyword{multivariate}