\name{panelVariables_en}
\alias{panelVariables_en}

\title{
Variables panel (English)
}
\description{
Options Variables selection panel. Select individuals and variables to be used in the analysis.
}
%\encoding{UTF-8}

\section{Selecting rows and columns}{
\itemize{
It consists of two independent lists: \emph{Select rows} and \emph{Select columns}, which have a list of the read elements. It is allowed selecting so as multiple options provided by the operating system. For Windows using press Shift and Control keys, in addition to the completed selection buttons.
  \item graphical buttons \emph{Remove all}: deselects the whole list.
  \item graphical buttons \emph{Check all}: selects the whole list.
  \item \emph{OK} button: Finishes the selection.
}
}
\section{General buttons}{
In the main program window some components are permanently visible.
\itemize{
  \item File loaded: reports data file is loaded.
  \item Button \emph{quit}: to close in an orderly, maintained in the environment files and format data and removes temporary files. It is always visible.
  \item Button \emph{Biplot}: for the biplot analysis. Set when marked enough options for analysis.
  \item Button \emph{Results}: to see the numerical results of the analysis. Is enabled after running the analysis.
}
}

\seealso{
  \code{\link{panelData_en}}, \code{\link{panelFormat_en}}, \code{\link{panelAnalysis_en}}.
}
\keyword{help}
