\name{panelVariables_es}
\alias{panelVariables_es}

\title{
Panel de Variables (Spanish)
}
\description{
Opciones del panel de selección de Variables. Permite seleccionar los individuos y las variables que se van a utilizar en el análisis.
}
\encoding{UTF-8}

\section{Selección de filas y columnas}{
\itemize{
Se compone de dos listas independientes: \emph{Selecciona filas} y \emph{Selecciona columnas}, con la lista de los elementos leídos. Se pueden seleccionar de forma múltiple según las opciones aportadas por el sistema operativo. En el caso de Windows con las teclas \emph{Mayúsculas} y \emph{Control}, además de los botones de selección completa.
  \item Botones gráficos \emph{Quita todas}: deselecciona toda la lista.
  \item Botones gráficos \emph{Marca todas}: selecciona toda la lista.
  \item Botón \emph{OK}: termina la selección.
}
}
\section{Botones generales}{
En la ventana principal del programa se encuentran algunos componentes visibles de forma permanente.
\itemize{
  \item Fichero cargado: informa del archivo desde el que se han cargado los datos.
  \item Botón \emph{Salir}: para cerrar de forma ordenada, mantiene en el entorno los ficheros de datos y de formato y elimina los ficheros temporales. Siempre está visible.
  \item Botón \emph{Biplot}: para efectuar el análisis biplot. Se habilita cuando se han marcado las opciones suficientes para realizar el análisis.
  \item Botón \emph{Resultados}: para ver los resultados numéricos del análisis. Se habilita una vez ejecutado el análisis.
}
}

\seealso{
  \code{\link{panelData_es}}, \code{\link{panelFormat_es}}, \code{\link{panelAnalysis_es}}.
}
\keyword{help}
